Ensure errors are caught.
Explanation: The rejection of a promise can only be caught if it is either awaited or returned.
The reason why is difficult to explain. What you want is that when a promise rejects, its parent promise also rejects. Its rejection only "bubbles up" in this way when the promise is awaited or returned.